+++ CHANGELOG.md
@@ -8,10 +8,9 @@
   [Olivier Halligon][AliSoftware]
   [#33](https://github.com/CocoaPods/CLAide/issues/33)
 
-* Support for argument alternatives + special ellipsis argument
+* Support for argument alternatives and repeatable arguments (ellipsis)
   [Olivier Halligon][AliSoftware]
   [#33](https://github.com/CocoaPods/CLAide/issues/33)
-
 
 ## 0.6.1